Extract SearchDoc out of DocExplorer, renaming it to SearchResults
Changed the name to obviate the need for the comment describing what th
component is for.
